Make `bbuf_get` result non-const, and add `bbuf_free`.
It's reasonable to want to modify the contents of the buffer in other
ways than just plain appending, and the storage is already guaranteed
contiguous. Drop the `const` from the `bbuf_get` return type, so it
can be used to modify bytes in-place.

The parallel request processing code may need to actually deallocate
buffers as well, so add a `bbuf_free` function.